refactor Intent & Commitment into mixin zome modules
no code changes, except struct field visibility
