+++ Cargo.toml
@@ -5,6 +5,11 @@ members= [
   "lib/vf_core",
   "lib/vf_planning",
   "lib/vf_knowledge",
+  "lib/rea_commitment/defs",
+  "lib/rea_commitment/lib",
+  "lib/rea_commitment/storage",
+  "lib/rea_commitment/storage_consts",
+  "lib/rea_commitment/rpc",
   "lib/rea_economic_event/defs",
   "lib/rea_economic_event/lib",
   "lib/rea_economic_event/structs_internal",
@@ -13,6 +18,11 @@ members= [
   "lib/rea_economic_resource/lib",
   "lib/rea_economic_resource/structs_internal",
   "lib/rea_economic_resource/structs_rpc",
+  "lib/rea_intent/defs",
+  "lib/rea_intent/lib",
+  "lib/rea_intent/storage",
+  "lib/rea_intent/storage_consts",
+  "lib/rea_intent/rpc",
   "lib/rea_process/defs",
   "lib/rea_process/lib",
   "lib/rea_process/structs_internal",
