+++ happs/observation/zomes/process/code/Cargo.toml
@@ -20,6 +20,8 @@ hc_zome_rea_process_storage_consts = { path = "../../../../../lib/rea_process/st
 hc_zome_rea_process_structs_internal = { path = "../../../../../lib/rea_process/structs_internal" }
 hc_zome_rea_process_structs_rpc = { path = "../../../../../lib/rea_process/structs_rpc" }
 hc_zome_rea_process_lib = { path = "../../../../../lib/rea_process/lib" }
+hc_zome_rea_commitment_storage_consts = { path = "../../../../../lib/rea_commitment/storage_consts" }
+hc_zome_rea_intent_storage_consts = { path = "../../../../../lib/rea_intent/storage_consts" }
 
 [lib]
 path = "src/lib.rs"
