+++ happs/planning/zomes/commitment/code/Cargo.toml
@@ -10,13 +10,13 @@ serde_json = { version = "=1.0.39", features = ["preserve_order"] }
 serde_derive = "=1.0.89"
 # :DUPE: hdk-rust-revid
 hdk = "=0.0.40-alpha1"
-hdk_proc_macros = "=0.0.40-alpha1"
-holochain_wasm_utils = "=0.0.40-alpha1"
-holochain_json_derive = "=0.0.17"
+hdk_proc_macros = "0.0.40-alpha1"
 
-hdk_graph_helpers = { path = "../../../../../lib/hdk_graph_helpers" }
-vf_observation = { path = "../../../../../lib/vf_observation" }
-vf_planning = { path = "../../../../../lib/vf_planning" }
+hc_zome_rea_commitment_defs = { path = "../../../../../lib/rea_commitment/defs" }
+hc_zome_rea_commitment_rpc = { path = "../../../../../lib/rea_commitment/rpc" }
+hc_zome_rea_commitment_lib = { path = "../../../../../lib/rea_commitment/lib" }
+hc_zome_rea_commitment_storage_consts = { path = "../../../../../lib/rea_commitment/storage_consts" }
+hc_zome_rea_intent_storage_consts = { path = "../../../../../lib/rea_intent/storage_consts" }
 hc_zome_rea_process_storage_consts = { path = "../../../../../lib/rea_process/storage_consts" }
 
 [lib]
