+++ happs/planning/zomes/commitment/code/src/lib.rs
@@ -13,55 +13,26 @@
  * @author:  pospi <pospi@spadgos.com>
  * @since:   2019-02-06
  */
-
-extern crate hdk;
 extern crate serde;
-#[macro_use]
-extern crate serde_derive;
-extern crate serde_json;
-extern crate hdk_graph_helpers;
-extern crate vf_planning;
-
-mod commitment_requests;
+extern crate hdk;
+extern crate hdk_proc_macros;
 
 use hdk::prelude::*;
 use hdk_proc_macros::zome;
 
-use vf_planning::type_aliases::CommitmentAddress;
-use vf_planning::commitment::{
-    Entry as CommitmentEntry,
-    CreateRequest as CommitmentCreateRequest,
-    UpdateRequest as CommitmentUpdateRequest,
-    ResponseData as CommitmentResponse,
-};
+use hc_zome_rea_commitment_defs::{ entry_def, base_entry_def };
+use hc_zome_rea_commitment_rpc::*;
+use hc_zome_rea_commitment_lib::*;
 
-use commitment_requests::{
-    QueryParams,
-    receive_get_commitment,
-    receive_create_commitment,
-    receive_update_commitment,
-    receive_delete_commitment,
-    receive_query_commitments,
-};
-
-use vf_planning::identifiers::{
-    COMMITMENT_BASE_ENTRY_TYPE,
-    COMMITMENT_INITIAL_ENTRY_LINK_TYPE,
-    COMMITMENT_ENTRY_TYPE,
-    COMMITMENT_FULFILLEDBY_LINK_TYPE,
-    FULFILLMENT_BASE_ENTRY_TYPE,
-    COMMITMENT_SATISFIES_LINK_TYPE,
-    SATISFACTION_BASE_ENTRY_TYPE,
-    COMMITMENT_INPUT_OF_LINK_TYPE,
-    COMMITMENT_OUTPUT_OF_LINK_TYPE,
-};
+// :TODO: split to own zome
+use hc_zome_rea_commitment_storage_consts::COMMITMENT_BASE_ENTRY_TYPE;
+use hc_zome_rea_intent_storage_consts::INTENT_BASE_ENTRY_TYPE;
 use hc_zome_rea_process_storage_consts::{
     PROCESS_BASE_ENTRY_TYPE,
-    PROCESS_COMMITMENT_INPUTS_LINK_TYPE, PROCESS_COMMITMENT_OUTPUTS_LINK_TYPE,
-    PROCESS_INTENT_INPUTS_LINK_TYPE, PROCESS_INTENT_OUTPUTS_LINK_TYPE,
-};
-use vf_planning::identifiers::{
-    INTENT_BASE_ENTRY_TYPE,
+    PROCESS_COMMITMENT_INPUTS_LINK_TYPE,
+    PROCESS_COMMITMENT_OUTPUTS_LINK_TYPE,
+    PROCESS_INTENT_INPUTS_LINK_TYPE,
+    PROCESS_INTENT_OUTPUTS_LINK_TYPE,
 };
 
 // Zome entry type wrappers
@@ -80,109 +51,12 @@ mod rea_commitment_planning_zome {
 
     #[entry_def]
     fn commitment_entry_def() -> ValidatingEntryType {
-        entry!(
-            name: COMMITMENT_ENTRY_TYPE,
-            description: "A planned economic flow that has been promised by an agent to another agent.",
-            sharing: Sharing::Public,
-            validation_package: || {
-                hdk::ValidationPackageDefinition::Entry
-            },
-            validation: |validation_data: hdk::EntryValidationData<CommitmentEntry>| {
-                // CREATE
-                if let EntryValidationData::Create{ entry, validation_data: _ } = validation_data {
-                    let record: CommitmentEntry = entry;
-                    let result = record.validate_or_fields();
-                    if result.is_ok() {
-                        return record.validate_action();
-                    }
-                    return result;
-                }
-
-                // UPDATE
-                if let EntryValidationData::Modify{ new_entry, old_entry: _, old_entry_header: _, validation_data: _ } = validation_data {
-                    let record: CommitmentEntry = new_entry;
-                    let result = record.validate_or_fields();
-                    if result.is_ok() {
-                        return record.validate_action();
-                    }
-                    return result;
-                }
-
-                // DELETE
-                // if let EntryValidationData::Delete{ old_entry, old_entry_header: _, validation_data: _ } = validation_data {
-
-                // }
-
-                Ok(())
-            }
-        )
+        entry_def()
     }
 
     #[entry_def]
     fn commitment_base_entry_def() -> ValidatingEntryType {
-        entry!(
-            name: COMMITMENT_BASE_ENTRY_TYPE,
-            description: "Base anchor for initial commitment addresses to provide lookup functionality",
-            sharing: Sharing::Public,
-            validation_package: || {
-                hdk::ValidationPackageDefinition::Entry
-            },
-            validation: |_validation_data: hdk::EntryValidationData<Address>| {
-                Ok(())
-            },
-            links: [
-                to!(
-                    COMMITMENT_ENTRY_TYPE,
-                    link_type: COMMITMENT_INITIAL_ENTRY_LINK_TYPE,
-                    validation_package: || {
-                        hdk::ValidationPackageDefinition::Entry
-                    },
-                    validation: | _validation_data: hdk::LinkValidationData| {
-                        Ok(())
-                    }
-                ),
-                to!(
-                    FULFILLMENT_BASE_ENTRY_TYPE,
-                    link_type: COMMITMENT_FULFILLEDBY_LINK_TYPE,
-                    validation_package: || {
-                        hdk::ValidationPackageDefinition::Entry
-                    },
-                    validation: | _validation_data: hdk::LinkValidationData| {
-                        Ok(())
-                    }
-                ),
-                to!(
-                    SATISFACTION_BASE_ENTRY_TYPE,
-                    link_type: COMMITMENT_SATISFIES_LINK_TYPE,
-                    validation_package: || {
-                        hdk::ValidationPackageDefinition::Entry
-                    },
-                    validation: | _validation_data: hdk::LinkValidationData| {
-                        Ok(())
-                    }
-                ),
-                to!(
-                    PROCESS_BASE_ENTRY_TYPE,
-                    link_type: COMMITMENT_INPUT_OF_LINK_TYPE,
-                    validation_package: || {
-                        hdk::ValidationPackageDefinition::Entry
-                    },
-                    validation: | _validation_data: hdk::LinkValidationData| {
-                        Ok(())
-                    }
-                ),
-                to!(
-                    PROCESS_BASE_ENTRY_TYPE,
-                    link_type: COMMITMENT_OUTPUT_OF_LINK_TYPE,
-                    validation_package: || {
-                        hdk::ValidationPackageDefinition::Entry
-                    },
-                    validation: | _validation_data: hdk::LinkValidationData| {
-                        Ok(())
-                    }
-                )
-            ]
-        )
+        base_entry_def()
     }
 
     #[entry_def]
@@ -247,17 +121,17 @@ mod rea_commitment_planning_zome {
     }
 
     #[zome_fn("hc_public")]
-    fn create_commitment(commitment: CommitmentCreateRequest) -> ZomeApiResult<CommitmentResponse>{
+    fn create_commitment(commitment: CreateRequest) -> ZomeApiResult<ResponseData> {
         receive_create_commitment(commitment)
     }
 
     #[zome_fn("hc_public")]
-    fn get_commitment(address: CommitmentAddress) -> ZomeApiResult<CommitmentResponse> {
+    fn get_commitment(address: CommitmentAddress) -> ZomeApiResult<ResponseData> {
         receive_get_commitment(address)
     }
 
     #[zome_fn("hc_public")]
-    fn update_commitment(commitment: CommitmentUpdateRequest) -> ZomeApiResult<CommitmentResponse> {
+    fn update_commitment(commitment: UpdateRequest) -> ZomeApiResult<ResponseData> {
         receive_update_commitment(commitment)
     }
 
@@ -267,7 +141,7 @@ mod rea_commitment_planning_zome {
     }
 
     #[zome_fn("hc_public")]
-    fn query_commitments(params: QueryParams ) -> ZomeApiResult<Vec<CommitmentResponse>> {
+    fn query_commitments(params: QueryParams) -> ZomeApiResult<Vec<ResponseData>>{
         receive_query_commitments(params)
     }
