+++ happs/planning/zomes/intent/code/Cargo.toml
@@ -10,13 +10,12 @@ serde_json = { version = "=1.0.39", features = ["preserve_order"] }
 serde_derive = "=1.0.89"
 # :DUPE: hdk-rust-revid
 hdk = "=0.0.40-alpha1"
-hdk_proc_macros = "=0.0.40-alpha1"
-holochain_wasm_utils = "=0.0.40-alpha1"
-holochain_json_derive = "=0.0.17"
+hdk_proc_macros = "0.0.40-alpha1"
 
-hdk_graph_helpers = { path = "../../../../../lib/hdk_graph_helpers" }
-vf_planning = { path = "../../../../../lib/vf_planning" }
-vf_observation = { path = "../../../../../lib/vf_observation" }
+hc_zome_rea_intent_defs = { path = "../../../../../lib/rea_intent/defs" }
+hc_zome_rea_intent_rpc = { path = "../../../../../lib/rea_intent/rpc" }
+hc_zome_rea_intent_lib = { path = "../../../../../lib/rea_intent/lib" }
+hc_zome_rea_intent_storage_consts = { path = "../../../../../lib/rea_intent/storage_consts" }
 hc_zome_rea_process_storage_consts = { path = "../../../../../lib/rea_process/storage_consts" }
 
 [lib]
