+++ happs/planning/zomes/satisfaction/code/Cargo.toml
@@ -16,6 +16,9 @@ holochain_json_derive = "=0.0.17"
 
 hdk_graph_helpers = { path = "../../../../../lib/hdk_graph_helpers" }
 vf_planning = { path = "../../../../../lib/vf_planning" }
+hc_zome_rea_commitment_rpc = { path = "../../../../../lib/rea_commitment/rpc" }
+hc_zome_rea_commitment_storage_consts = { path = "../../../../../lib/rea_commitment/storage_consts" }
+hc_zome_rea_intent_storage_consts = { path = "../../../../../lib/rea_intent/storage_consts" }
 
 [lib]
 path = "src/lib.rs"
