+++ lib/rea_commitment/lib/Cargo.toml
@@ -1,0 +1,20 @@
+[package]
+name = "hc_zome_rea_commitment_lib"
+version = "0.1.0"
+authors = ["pospi <pospi@spadgos.com>"]
+edition = "2018"
+
+[dependencies]
+# :DUPE: hdk-rust-revid
+hdk = "=0.0.40-alpha1"
+
+hdk_graph_helpers = { path = "../../hdk_graph_helpers" }
+vf_core = { path = "../../vf_core" }
+vf_planning = { path = "../../vf_planning" }
+hc_zome_rea_commitment_storage_consts = { path = "../storage_consts" }
+hc_zome_rea_commitment_storage = { path = "../storage" }
+hc_zome_rea_commitment_rpc = { path = "../rpc" }
+hc_zome_rea_process_storage_consts = { path = "../../rea_process/storage_consts" }
+
+[lib]
+crate-type = ["lib"]
