+++ lib/rea_intent/defs/Cargo.toml
@@ -1,0 +1,23 @@
+[package]
+name = "hc_zome_rea_intent_defs"
+version = "0.1.0"
+authors = ["pospi <pospi@spadgos.com>"]
+edition = "2018"
+
+[dependencies]
+serde = "1.0.89"
+serde_json = { version = "=1.0.39", features = ["preserve_order"] }
+serde_derive = "=1.0.89"
+# :DUPE: hdk-rust-revid
+hdk = "=0.0.40-alpha1"
+holochain_json_derive = "=0.0.17"
+
+hdk_graph_helpers = { path = "../../hdk_graph_helpers" }
+vf_planning = { path = "../../vf_planning" }
+hc_zome_rea_intent_storage = { path = "../storage" }
+hc_zome_rea_intent_storage_consts = { path = "../storage_consts" }
+hc_zome_rea_intent_rpc = { path = "../rpc" }
+hc_zome_rea_process_storage_consts = { path = "../../rea_process/storage_consts" }
+
+[lib]
+crate-type = ["lib"]
