+++ lib/rea_intent/lib/src/lib.rs
@@ -1,13 +1,17 @@
 /**
- * Handling for `Intent`-related requests
+ * Holo-REA intent zome library API
+ *
+ * Contains helper methods that can be used to manipulate `Intent` data
+ * structures in either the local Holochain zome, or a separate DNA-local zome.
+ *
+ * @package Holo-REA
  */
+use std::borrow::Cow;
 use hdk::{
     PUBLIC_TOKEN,
-    holochain_persistence_api::cas::content::Address,
-    holochain_json_api::{ json::JsonString, error::JsonError },
+    prelude::Address,
     error::{ ZomeApiResult, ZomeApiError },
 };
-use holochain_json_derive::{ DefaultJson };
 
 use hdk_graph_helpers::{
     MaybeUndefined,
@@ -16,6 +20,9 @@ use hdk_graph_helpers::{
         read_record_entry,
         update_record,
         delete_record,
+    },
+    links::{
+        get_linked_addresses_as_type,
     },
     local_indexes::{
         query_direct_index_with_foreign_key,
@@ -28,53 +35,33 @@ use hdk_graph_helpers::{
     },
 };
 
-use vf_observation::type_aliases::{
-    ProcessAddress,
-};
-use vf_planning::type_aliases::{
-    IntentAddress,
+use vf_core::type_aliases::{
     SatisfactionAddress,
 };
-use vf_planning::identifiers::{
-    BRIDGED_OBSERVATION_DHT,
-    INTENT_BASE_ENTRY_TYPE,
-    INTENT_INITIAL_ENTRY_LINK_TYPE,
-    INTENT_ENTRY_TYPE,
-    INTENT_INPUT_OF_LINK_TYPE, INTENT_INPUT_OF_LINK_TAG,
-    INTENT_OUTPUT_OF_LINK_TYPE, INTENT_OUTPUT_OF_LINK_TAG,
-    SATISFACTION_SATISFIES_LINK_TYPE, SATISFACTION_SATISFIES_LINK_TAG,
-};
+
+use hc_zome_rea_intent_storage_consts::*;
+use hc_zome_rea_intent_storage::*;
+use hc_zome_rea_intent_rpc::*;
+
 use hc_zome_rea_process_storage_consts::{
     PROCESS_BASE_ENTRY_TYPE,
     PROCESS_INTENT_INPUTS_LINK_TYPE, PROCESS_INTENT_INPUTS_LINK_TAG,
     PROCESS_INTENT_OUTPUTS_LINK_TYPE, PROCESS_INTENT_OUTPUTS_LINK_TAG,
 };
-use vf_planning::intent::{
-    Entry,
-    CreateRequest,
-    UpdateRequest,
-    ResponseData as Response,
-    construct_response,
-    get_link_fields,
+
+use vf_planning::identifiers::{
+    SATISFACTION_SATISFIES_LINK_TYPE, SATISFACTION_SATISFIES_LINK_TAG,
 };
 
-#[derive(Serialize, Deserialize, Debug, DefaultJson, Clone)]
-#[serde(rename_all = "camelCase")]
-pub struct QueryParams {
-    input_of: Option<ProcessAddress>,
-    output_of: Option<ProcessAddress>,
-    satisfied_by: Option<SatisfactionAddress>,
-}
-
-pub fn receive_create_intent(intent: CreateRequest) -> ZomeApiResult<Response> {
+pub fn receive_create_intent(intent: CreateRequest) -> ZomeApiResult<ResponseData> {
     handle_create_intent(&intent)
 }
 
-pub fn receive_get_intent(address: IntentAddress) -> ZomeApiResult<Response> {
+pub fn receive_get_intent(address: IntentAddress) -> ZomeApiResult<ResponseData> {
     handle_get_intent(&address)
 }
 
-pub fn receive_update_intent(intent: UpdateRequest) -> ZomeApiResult<Response> {
+pub fn receive_update_intent(intent: UpdateRequest) -> ZomeApiResult<ResponseData> {
     handle_update_intent(&intent)
 }
 
@@ -82,18 +69,18 @@ pub fn receive_delete_intent(address: IntentAddress) -> ZomeApiResult<bool> {
     handle_delete_intent(&address)
 }
 
-pub fn receive_query_intents(params: QueryParams) -> ZomeApiResult<Vec<Response>> {
+pub fn receive_query_intents(params: QueryParams) -> ZomeApiResult<Vec<ResponseData>> {
     handle_query_intents(&params)
 }
 
 // :TODO: move to hdk_graph_helpers module
 
-fn handle_get_intent(address: &IntentAddress) -> ZomeApiResult<Response> {
+fn handle_get_intent(address: &IntentAddress) -> ZomeApiResult<ResponseData> {
     let entry = read_record_entry(&address)?;
     Ok(construct_response(&address, &entry, get_link_fields(&address)))
 }
 
-fn handle_create_intent(intent: &CreateRequest) -> ZomeApiResult<Response> {
+fn handle_create_intent(intent: &CreateRequest) -> ZomeApiResult<ResponseData> {
     let (base_address, entry_resp): (IntentAddress, Entry) = create_record(
         INTENT_BASE_ENTRY_TYPE, INTENT_ENTRY_TYPE,
         INTENT_INITIAL_ENTRY_LINK_TYPE,
@@ -126,7 +113,7 @@ fn handle_create_intent(intent: &CreateRequest) -> ZomeApiResult<Response> {
     Ok(construct_response(&base_address, &entry_resp, get_link_fields(&base_address)))
 }
 
-fn handle_update_intent(intent: &UpdateRequest) -> ZomeApiResult<Response> {
+fn handle_update_intent(intent: &UpdateRequest) -> ZomeApiResult<ResponseData> {
     let address = intent.get_id();
     let new_entry = update_record(INTENT_ENTRY_TYPE, address, intent)?;
 
@@ -182,7 +169,7 @@ fn handle_delete_intent(address: &IntentAddress) -> ZomeApiResult<bool> {
     delete_record::<Entry>(&address)
 }
 
-fn handle_query_intents(params: &QueryParams) -> ZomeApiResult<Vec<Response>> {
+fn handle_query_intents(params: &QueryParams) -> ZomeApiResult<Vec<ResponseData>> {
     let mut entries_result: ZomeApiResult<Vec<(IntentAddress, Option<Entry>)>> = Err(ZomeApiError::Internal("No results found".to_string()));
 
     match &params.satisfied_by {
@@ -228,4 +215,55 @@ fn handle_query_intents(params: &QueryParams) -> ZomeApiResult<Vec<Response>> {
         ),
         _ => Err(ZomeApiError::Internal("could not load linked addresses".to_string()))
     }
+}
+
+/// Create response from input DHT primitives
+pub fn construct_response<'a>(
+    address: &IntentAddress, e: &Entry, (
+        satisfactions,
+        // published_in,
+    ): (
+        Option<Cow<'a, Vec<SatisfactionAddress>>>,
+        // Option<Cow<'a, Vec<ProposedIntentAddress>>>
+    )
+) -> ResponseData {
+    ResponseData {
+        intent: Response {
+            id: address.to_owned(),
+            action: e.action.to_owned(),
+            note: e.note.to_owned(),
+            image: e.image.to_owned(),
+            input_of: e.input_of.to_owned(),
+            output_of: e.output_of.to_owned(),
+            provider: e.provider.to_owned(),
+            receiver: e.receiver.to_owned(),
+            resource_inventoried_as: e.resource_inventoried_as.to_owned(),
+            resource_classified_as: e.resource_classified_as.to_owned(),
+            resource_conforms_to: e.resource_conforms_to.to_owned(),
+            resource_quantity: e.resource_quantity.to_owned(),
+            effort_quantity: e.effort_quantity.to_owned(),
+            available_quantity: e.available_quantity.to_owned(),
+            has_beginning: e.has_beginning.to_owned(),
+            has_end: e.has_end.to_owned(),
+            has_point_in_time: e.has_point_in_time.to_owned(),
+            due: e.due.to_owned(),
+            at_location: e.at_location.to_owned(),
+            agreed_in: e.agreed_in.to_owned(),
+            finished: e.finished.to_owned(),
+            in_scope_of: e.in_scope_of.to_owned(),
+            satisfied_by: satisfactions.map(Cow::into_owned),
+            // published_in: published_in.map(Cow::into_owned),
+        }
+    }
+}
+
+//---------------- READ ----------------
+
+// @see construct_response
+pub fn get_link_fields<'a>(intent: &IntentAddress) -> (
+    Option<Cow<'a, Vec<SatisfactionAddress>>>,
+) {
+    (
+        Some(get_linked_addresses_as_type(intent, INTENT_SATISFIEDBY_LINK_TYPE, INTENT_SATISFIEDBY_LINK_TAG)),
+    )
 }
