+++ lib/rea_process/lib/Cargo.toml
@@ -15,6 +15,8 @@ hc_zome_rea_process_storage_consts = { path = "../storage_consts" }
 hc_zome_rea_process_structs_internal = { path = "../structs_internal" }
 hc_zome_rea_process_structs_rpc = { path = "../structs_rpc" }
 hc_zome_rea_economic_event_storage_consts = { path = "../../rea_economic_event/storage_consts" }
+hc_zome_rea_commitment_storage_consts = { path = "../../rea_commitment/storage_consts" }
+hc_zome_rea_intent_storage_consts = { path = "../../rea_intent/storage_consts" }
 
 [lib]
 crate-type = ["lib"]
