+++ lib/vf_planning/src/lib.rs
@@ -12,25 +12,11 @@ extern crate vf_core;
 
 pub use vf_core::{ type_aliases, measurement };
 
-pub mod commitment;
 pub mod fulfillment;
-pub mod intent;
 pub mod satisfaction;
 
 pub mod identifiers {
     pub const BRIDGED_OBSERVATION_DHT: &str = "vf_observation";
-
-    pub const COMMITMENT_BASE_ENTRY_TYPE: &str = "vf_commitment_baseurl";
-    pub const COMMITMENT_INITIAL_ENTRY_LINK_TYPE: &str = "vf_commitment_entry";
-    pub const COMMITMENT_ENTRY_TYPE: &str = "vf_commitment";
-    pub const COMMITMENT_FULFILLEDBY_LINK_TYPE: &str = "vf_commitment_fulfilled_by";
-    pub const COMMITMENT_FULFILLEDBY_LINK_TAG: &str = "fulfills";
-    pub const COMMITMENT_SATISFIES_LINK_TYPE: &str = "vf_commitment_satisfies";
-    pub const COMMITMENT_SATISFIES_LINK_TAG: &str = "satisfies";
-    pub const COMMITMENT_INPUT_OF_LINK_TYPE: &str = "vf_commitment_input_of";
-    pub const COMMITMENT_INPUT_OF_LINK_TAG: &str = "input_of";
-    pub const COMMITMENT_OUTPUT_OF_LINK_TYPE: &str = "vf_commitment_output_of";
-    pub const COMMITMENT_OUTPUT_OF_LINK_TAG: &str = "output_of";
 
     pub const FULFILLMENT_BASE_ENTRY_TYPE: &str = "vf_fulfillment_baseurl";
     pub const FULFILLMENT_INITIAL_ENTRY_LINK_TYPE: &str = "vf_fulfillment_entry";
@@ -39,16 +25,6 @@ pub mod identifiers {
     pub const FULFILLMENT_FULFILLS_LINK_TAG: &str = "fulfills";
     pub const FULFILLMENT_FULFILLEDBY_LINK_TYPE: &str = "vf_fulfillment_fulfilled_by";
     pub const FULFILLMENT_FULFILLEDBY_LINK_TAG: &str = "fulfilled_by";
-
-    pub const INTENT_BASE_ENTRY_TYPE: &str = "vf_intent_baseurl";
-    pub const INTENT_INITIAL_ENTRY_LINK_TYPE: &str = "vf_intent_entry";
-    pub const INTENT_ENTRY_TYPE: &str = "vf_intent";
-    pub const INTENT_SATISFIEDBY_LINK_TYPE: &str = "vf_intent_satisfied_by";
-    pub const INTENT_SATISFIEDBY_LINK_TAG: &str = "satisfied_by";
-    pub const INTENT_INPUT_OF_LINK_TYPE: &str = "vf_intent_input_of";
-    pub const INTENT_INPUT_OF_LINK_TAG: &str = "input_of";
-    pub const INTENT_OUTPUT_OF_LINK_TYPE: &str = "vf_intent_output_of";
-    pub const INTENT_OUTPUT_OF_LINK_TAG: &str = "output_of";
 
     pub const SATISFACTION_BASE_ENTRY_TYPE: &str = "vf_satisfaction_baseurl";
     pub const SATISFACTION_INITIAL_ENTRY_LINK_TYPE: &str = "vf_satisfaction_entry";
